pyproject.toml
[project]
name = "MCGram"
version = "0.4.0"
description = "Telegarm robot, query the official genshin information."
authors = [
    {name = "洛水居室"},
    {name = "zhxy-CN"},
    {name = "Chuangbo Li"},
    {name = "kotoriのねこ"},
    {name = "omg-xtao"},
    {name = "艾迪"},
    {name = "Karako"},
    {name = "SiHuaN"},
]
dependencies = [
    "httpx<1.0.0,>=0.25.0",
    "ujson<6.0.0,>=5.9.0",
    "Jinja2<4.0.0,>=3.1.2",
    "python-telegram-bot[ext,rate-limiter]<22.0,>=21.3",
    "sqlmodel<1.0.0,>=0.0.16",
    "colorlog<7.0.0,>=6.8.0",
    "fakeredis<3.0.0,>=2.19.0",
    "redis<6.0.0,>=5.0.1",
    "beautifulsoup4<5.0.0,>=4.12.1",
    "asyncmy<1.0.0,>=0.2.9",
    "aiofiles<24.0.0,>=23.2.1",
    "python-dotenv<2.0.0,>=1.0.0",
    "alembic<2.0.0,>=1.13.0",
    "black<25.0.0,>=24.1.1",
    "rich<14.0.0,>=13.6.0",
    "enkanetwork-py @ git+https://github.com/PaiGramTeam/EnkaNetwork.py",
    "lxml<6.0.0,>=5.0.0",
    "arko-wrapper<1.0.0,>=0.2.8",
    "fastapi<1.0.0,>=0.110.0",
    "uvicorn[standard]<1.0.0,>=0.28.0",
    "sentry-sdk<2.0.0,>=1.31.0",
    "GitPython<4.0.0,>=3.1.30",
    "openpyxl<4.0.0,>=3.1.1",
    "async-lru<3.0.0,>=2.0.4",
    "thefuzz<1.0.0,>=0.22.1",
    "cryptography<43.0.0,>=42.0.2",
    "pillow<11.0.0,>=10.0.1",
    "playwright==1.39.0",
    "aiosqlite[sqlite]<1.0.0,>=0.20.0",
    "simnet @ git+https://github.com/PaiGramTeam/SIMNet",
    "gcsim-pypi<3.0.0.0,>=2.8.2.10",
    "psutil<6.0.0,>=5.9.6",
]
requires-python = "<4.0,>=3.8"
readme = "README.md"
license = {text = "AGPL-3.0"}

[project.optional-dependencies]
pyro = ["TgCrypto<2.0.0,>=1.2.5", "Pyrogram<3.0.0,>=2.0.102"]
test = ["pytest<8.0.0,>=7.3.0", "pytest-asyncio<1.0.0,>=0.23.2", "flaky<4.0.0,>=3.7.0"]
genshin-artifact = ["python-genshin-artifact<2.0.0,>=1.0.4"]

[tool.pdm]
distribution = false

[tool.pdm.build]
includes = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

# 格式配置
[tool.black]
include = '\.pyi?$'
line-length = 120
target-version = ['py311']
